feat: add exports to support bundlers targeting platforms like Cloudflare
Without this we are forced to import from "faunadb/dist/faunadb" and this sucks when working in TS as there are no types associated with the deep import.

Adding a default field to exports that points to this instead allows us to resolve the proper version when targeting platforms that don't "require()" the module, but rather "import" it.